rdma-core-17

Updates from version 16:

- debian: Add Debian uploads up to version 16.2-1
- debian: Override intentional WantedBy lintian warning
- debian: Guard udevadm call again
- vmw_pvrdma: Drop the extraneous align_next_power2 for sges
- Check that published headers do not rely on internal headers
- Always use internal kernel headers when building
- Support distros that have old kernel uapi headers
- rxe: Do not use _sockaddr in struct rxe_av
- rxe: Remove duplicate include
- verbs: Do not export kern-abi.h as a public header
- Fix compiler warnings on PPC64 and others
- mlx4: Fix 1<<31 expressions
- Enable GNU99 mode if the compiler does not support C11
- vmw_pvrdma: Fix SRQN returned to library
- mlx5: Add missing include file in mlx5dv.h
- redhat: remove files that we no longer use
- Don't include ifup/ifdown-ib
- build: Pass extra cmake flags
- iwpmd: fix double mutex unlock
- libbnxt_re: Update and sort device pci-id table
- libbnxt_re: Display FW version using ibv_devinfo
- libbnxt_re: Fix lat test failure in event mode
- cbuild: Replace Fedora 26 with Fedora 27
- rc_pingpong: Fix getopt processing of -t
- verbs: Use correct type for printf from struct rlimit
- librdmacm: Fix off-by-one errors in sscanf
- libibumad: Fix memory leak in umad_sa_mcm_rereg_test
- librdmacm: Set errno correctly if status is positive
- mlx5: Destroy and modify a DC target QP
- mlx5: Create DC transport QPs
- mlx5: Add interface to create DC transport QPs
- verbs: Introduce driver QP type
- ibacm: Close file descriptor in case of out-of-memory
- ibacm: Fix off-by-one errors in scanf for opt and value
- ibacm: Fix off-by-one errors in scanf for dev and pkey_str
- verbs: Add support to read completion's timestamp in wall clock nanoseconds
- mlx5: Add direct verb to translate device timestamp to nanoseconds
- mlx5: Add direct verbs support for query clock info
- mlx5: Support for user space clock info
- verbs: Tidy up the remaining structs in kern-abi.h
- verb: Use kernel uapi header in struct ex_hdr
- verbs: Use kernel uapi header for ibv_modify_qp_common
- verbs: Use kernel uapi header for flow steering types
- verbs: Use kernel uapi header for IBV_CREATE_QP_EX_KERNEL_MASK_IND_TABLE
- verbs: Use kernel uapi header for ibv_kern_send_wr
- verbs: Use kernel uapi header for ibv_create_ah_resp
- verbs: Use kernel uapi header for ibv_kern_wc and ibv_poll_cq_resp
- verbs: Use kernel uapi header for ibv_query_device_resp_ex
- verbs: Use kernel uapi header for ibv_cq_moderation_caps_resp
- verbs: Use kernel uapi names for identical structs
- verbs: Remove tests for NULL ops
- verbs: Add missing endian annotations to kern-abi.h
- verbs: Convert all providers to use verbs_set_ops
- verbs: Remove kernel ABI v2 compat cruft
- verbs: Provide a default implementation for every verbs op
- verbs: Remove unused structs
- verbs: Remove init_context/uninit_context
- verbs: Delete extended command definitions from kern-abi.h
- mlx5: Convert from init_context to alloc_context
- verbs: Include rdma/ib_user_verbs.h in kern-abi.h
- mlx4: Convert from init_context to alloc_context
- Update rdma/ib_user_verbs.h
- bnxt_re: Convert from init_context to alloc_context
- verbs: Always allocate a verbs_context
- debian: Do not require valgrind on ia64
- travis: Use clang-5.0
- travis: Do not loose ninja output if ninja fails during sparse
- cbuild: Support distros that only ship python3
- debian: Update changelog to include Debian uploads 16.x
- debian: Bump Standards-Version to 4.1.3
- debian: Add udev dependency to rdma-core and srptools
- verbs: Remove comp_mask checks from verbs.h
- mlx5: Expose max dynamic blue flame registers capability
- mlx5: Add support for thread domain as part of QP creation
- mlx5: Add support for using a dedicated BF register
- mlx5: Add support for dynamic UAR allocation
- mlx5: Dynamic BF size handling
- mlx5: Add support for ibv_parent_domain
- mlx5: Add support for the ibv_td object and its related verbs
- Remove the obsolete libibcm library
- verbs: Introduce ibv_parent_domain and its related verbs
- verbs: Introduce ibv_thread_domain and its related verbs
- Fix typos 'pontentially' and 'bandwith'
- srp_daemon: Don't create async_ev_thread if only run once
- srp_daemon: Remove duplicate sync_res shutdown
- srp_daemon: fix CQ handling
- srp_daemon: handle SM lid change
- Have cbuild map the git alternates
- providers: Remove unneeded function prototypes
- hns: Remove useless assignments
- mlx4: Don't nest container_of
- verbs: Remove container_of from the public verbs.h
- verbs: Don't open code 'container_of' in the providers
- vmw_pvrdma: Remove usage of BIT() from vmw_pvrdma
- libibverb/examples: Add command line option to enable buffer validation
- register_to_trap(): Specify number of iterations once
- srp_daemon.c: Cosmetic change to some error messages
- verbs: Get rid of verbs_set_ctx_op
- verbs: Remove dead _HAS_ defines for providers
- udev.md: Fix typo 'le.g.acy' -> 'legacy'

+++ libibverbs/init.c
@@ -552,9 +552,9 @@ static void check_memlock_limit(void)
 	}
 
 	if (rlim.rlim_cur <= 32768)
-		fprintf(stderr, PFX "Warning: RLIMIT_MEMLOCK is %lu bytes.\n"
+		fprintf(stderr, PFX "Warning: RLIMIT_MEMLOCK is %llu bytes.\n"
 			"    This will severely limit memory registrations.\n",
-			rlim.rlim_cur);
+			(unsigned long long)rlim.rlim_cur);
 }
 
 static int same_sysfs_dev(struct verbs_sysfs_dev *sysfs1,
